requirements.txt
cryptography
python-dotenv
keyring